FEM: Fenics meshes, add support for xdmf format, some more improvements
