+++ src/Mod/Fem/App/CMakeLists.txt
@@ -72,6 +72,10 @@ SET(FemScripts_SRCS
     convert2TetGen.py
     importCcxDatResults.py
     importCcxFrdResults.py
+    readFenicsXML.py
+    writeFenicsXML.py
+    readFenicsXDMF.py
+    writeFenicsXDMF.py
     importFenicsMesh.py
     importInpMesh.py
     importToolsFem.py
