+++ src/Mod/Fem/CMakeLists.txt
@@ -20,7 +20,11 @@ INSTALL(
         convert2TetGen.py
         importCcxDatResults.py
         importCcxFrdResults.py
-        importFenicsMesh.py
+	    readFenicsXML.py
+	    writeFenicsXML.py
+	    readFenicsXDMF.py
+	    writeFenicsXDMF.py
+	    importFenicsMesh.py
         importInpMesh.py
         importToolsFem.py
         importVTKResults.py
